Fixed and tested game loop networking
Made a few changes to the server/client parts of the game loop as well as the server/client files themselves. Tested and successfully connected remotely as well as sent a test message both ways. Still unable to truly test sending moves back and forth since the join menu isn't tied in yet.